requirements.txt
# MCP Server
mcp>=1.0.0

# Data
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0

# Technical Analysis
ta>=0.11.0          # TA-lib wrapper (BB, RSI, MACD, SMA, EMA)

# LLM
anthropic>=0.30.0

# Async / utils
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# EDGAR
edgar>=5.4.0         # python-edgar wrapper